moved mapping client context
