+++ packages/itwin/grouping-mapping-widget/src/grouping-mapping-widget.ts
@@ -6,7 +6,7 @@
 export * from "./widget/GroupingMappingWidget";
 
 /** Interfaces for providing custom MappingClient */
-export { createDefaultMappingClient } from "./widget/components/GroupingMapping";
+export { createDefaultMappingClient } from "./widget/components/context/MappingClientContext";
 export { ClientPrefix } from "./widget/components/context/GroupingApiConfigContext";
 export * from "@itwin/insights-client";
 export * from "./widget/IMappingClient";
