+++ packages/itwin/grouping-mapping-widget/src/widget/components/CalculatedPropertyTable.tsx
@@ -21,7 +21,7 @@ import { PropertyMenuView } from "./PropertyMenu";
 import type { CellProps } from "react-table";
 import DeleteModal from "./DeleteModal";
 import type { CalculatedProperty } from "@itwin/insights-client";
-import { MappingClientContext } from "./GroupingMapping";
+import { useMappingClient } from "./context/MappingClientContext";
 import { useGroupingMappingApiConfig } from "./context/GroupingApiConfigContext";
 
 export type CalculatedPropertyType =
@@ -53,7 +53,7 @@ const CalculatedPropertyTable = ({
   selectedCalculatedProperty,
 }: CalculatedPropertyTableProps) => {
   const apiContext = useGroupingMappingApiConfig();
-  const mappingClient = useContext(MappingClientContext);
+  const mappingClient = useMappingClient();
   const [
     showCalculatedPropertyDeleteModal,
     setShowCalculatedPropertyDeleteModal,
